ADD RDS based PDS Nucleus product completion checker lambda functions
